requirements.txt
fastapi 
uvicorn 
python-multipart 
pillow 
replicate 
gradio
torch